+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 14. Updated Cpp2IL to v2021.0.
 15. Swapped out Il2CppDumper for Cpp2IL.
 16. Added "--melonloader.agfoffline" Launch Option to Force the Assembly Generator to Run without Contacting the Remote API.
+17. Fixed Issue with MelonLogger not abiding by Warning and Error Caps.
 
 ---